+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -2,7 +2,7 @@ package org.apache.coyote.http11;
 
 import org.apache.coyote.Processor;
 import org.apache.catalina.controller.Controller;
-import org.apache.catalina.controller.RequestMapping;
+import org.apache.catalina.RequestMapping;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import java.io.InputStream;
