+++ tomcat/src/test/java/org/apache/catalina/controller/LoginControllerTest.java
@@ -1,5 +1,6 @@
 package org.apache.catalina.controller;
 
+import com.techcourse.controller.LoginController;
 import org.apache.catalina.SessionManager;
 import org.apache.coyote.http11.HttpRequestHeaders;
 import org.apache.coyote.http11.HttpRequest;
